feat: add shared storage module
